Create custAndOrderInfoNext_action
Here's some basic VBA I wrote to save user inputted data from the form I created in to the actual Excel sheet. This is some of my earlier work so I wouldn't say it's my best, but it gets the job done. Next time I'd create different modules in order to separate the data from the actual functions (as in proper coding technique, but hey I was new). There's just a lot of boiler plate stuff that could be sorted better.